Zenyx BETA/requirements.txt
python-telegram-bot==13.15
redis==4.5.4
mysql-connector-python==8.0.33
requests==2.29.0
python-dotenv==1.0.0
Pillow==9.5.0